+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Policy management
 - Latest telemetry data for given device & type
+- Assignment and unassignment policy to device
 
 ## [1.7.1] - 2021-12-17
 ### Fixed